+++ pcs/lib/commands/constraint/colocation.py
@@ -4,8 +4,8 @@ from pcs.lib.cib.constraint import colocation
 import pcs.lib.commands.constraint.common
 
 # configure common constraint command
-show = partial(
-    pcs.lib.commands.constraint.common.show,
+config = partial(
+    pcs.lib.commands.constraint.common.config,
     colocation.TAG_NAME,
     lambda element: element.attrib.has_key("rsc"),
 )